+++ EmployeeManagementSystem/WebApp/DemoControllers/EmployeesController.cs
@@ -6,7 +6,7 @@ namespace WebApp.DemoControllers
 
 	public class EmployeesController : Controller
 	{
-		public IActionResult GetEmployeesByDepartment([FromRoute(Name = "Id")] int departmentId)
+		public IActionResult GetEmployeesByDepartment([FromRoute(Name = "Id")] long departmentId)
 		{
 			return Content($"Loading employees under department: {departmentId}");
 		}
